requirements.txt
webdriver_manager
selenium==4.8.2
openpyxl
python-decouple
pandas
redmail
chromedriver-autoinstaller 
